Create README.md
This Repository is created for Learning and Training Purposes.

For the beginners in Web Technologies, these samples are very effective when they want to create UI using HTML, CSS, Angular, Bootstrap, JS etc.

Apart from this, in these samples we take care of Responsive UI should be prepared as the new world is totally flexible in terms of using platforms for viewing the websites. It should be compatible to all platforms like Windows, Mac, Linux, Android, iOS etc.